Rename commit model to commit diff
